web3-market-data-stream-lit/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
aiohttp==3.9.3
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
